Task 3 - Hot fix

+++ 0x0A-argc_argv/3-mul.c
@@ -27,7 +27,7 @@ int main(int argc, char *argv[])
 	{
 		for (i = 1; i < argc; i++)
 		{
-			Result *= atoi(argv[i]);
+			Result *= atol(argv[i]);
 		}
 
 	}
